+++ R/inspectMetaDifferences.R
@@ -54,20 +54,12 @@ inspectMetaDifferences <- function(GADSdat, varName, other_GADSdat = GADSdat, ot
 
   varDiff <- "all.equal"
   if(!identical(metaVar1, metaVar2)) {
-    # only return column with differences
-    if(!identical(metaVar1$varLabel, metaVar2$varLabel)) {
-      varDiff <- data.frame(metaVar1$varLabel, metaVar2$varLabel)
-      names(varDiff) <- paste(nam_col, "varLabel", sep = "_")
-      if(!identical(metaVar1$format, metaVar2$format)) {
-        varDiff2 <- data.frame(metaVar1$format, metaVar2$format)
-        names(varDiff2) <- paste(nam_col, "format", sep = "_")
+    varLabel_diff <- data.frame(metaVar1$varLabel, metaVar2$varLabel)
+    names(varLabel_diff) <- paste(nam_col, "varLabel", sep = "_")
+    format_diff <- data.frame(metaVar1$format, metaVar2$format)
+    names(format_diff) <- paste(nam_col, "format", sep = "_")
 
-        varDiff <- cbind(varDiff, varDiff2)[, c(1, 3, 2, 4)]
-      }
-    } else {
-      varDiff <- data.frame(metaVar1$format, metaVar2$format)
-      names(varDiff) <- paste(nam_col, "format", sep = "_")
-    }
+    varDiff <- cbind(varLabel_diff, format_diff)[, c(1, 3, 2, 4)]
   }
 
   ## Value level